requirements.txt
# for this project
pytest
python-dotenv
